Move macOS dSYM test to its own directory
